backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==4.0.1
flask-migrate==4.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
gunicorn==22.0.0
reportlab==4.2.2
Pillow==10.4.0
google-generativeai==0.7.2
PyMuPDF==1.24.5
